Fixed a lot
I fixed the player movement, deleted the lanes, fixed the player collision and added a collision effect, also fixed the obstacle spawner and obstacles to work with our current lane layout and be randomized properly. Made it so that the collision could change image and have a sound effect if we want and also set up the player collision for a possible effect(Minus points or lives) if we want to add that later.