Did a lot . . .
Added in coins using the same basic code as the obstacles.
Added collision for coins that add a static point value to the score when you collect a coin.
Added lived to the Player.
Added end scenario when there are no lives left which drops to main menu.
All are very rough and need a second pass, but preliminary programming is done.